chore: initial project setup - remove browser opening for storybook - add dark mode to storybook - add next-themes and light/dark mode switcher to header - replace existing sort imports with simple-import-sort in eslint - pin node version with volta
- add ShopTemplate with header/footer
- add initial components and blocks
- move demo routes to /marketing
- add project root route and /product/:id route
