+++ .storybook/main.ts
@@ -7,6 +7,8 @@ const config: StorybookConfig = {
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/addon-interactions',
+    '@storybook/addon-controls',
+    'storybook-dark-mode',
   ],
   framework: {
     name: '@storybook/nextjs',
