+++ .storybook/preview.ts
@@ -1,5 +1,7 @@
-import type { Preview } from '@storybook/react';
 import '../src/styles/global.css';
+
+import type { Preview } from '@storybook/react';
+import { themes } from '@storybook/theming';
 
 const preview: Preview = {
   parameters: {
@@ -11,6 +13,15 @@ const preview: Preview = {
     },
     nextjs: {
       appDirectory: true,
+    },
+    darkMode: {
+      current: 'dark',
+      // Override the default dark theme
+      dark: { ...themes.dark, appBg: 'black' },
+      // Override the default light theme
+      light: { ...themes.normal, appBg: 'white' },
+      classTarget: 'html',
+      stylePreview: true,
     },
   },
 };
