+++ package.json
@@ -19,7 +19,7 @@
     "db:generate": "drizzle-kit generate",
     "db:migrate": "dotenv -c production -- drizzle-kit migrate",
     "db:studio": "dotenv -c production -- drizzle-kit studio",
-    "storybook": "storybook dev -p 6006",
+    "storybook": "storybook dev -p 6006 --no-open",
     "storybook:build": "storybook build",
     "storybook:serve": "http-server storybook-static --port 6006 --silent",
     "serve-storybook": "run-s storybook:*",
@@ -38,6 +38,7 @@
     "@radix-ui/react-tabs": "^1.1.1",
     "@sentry/nextjs": "^8.41.0",
     "@spotlightjs/spotlight": "^2.6.3",
+    "@storybook/addon-controls": "^8.4.7",
     "@t3-oss/env-nextjs": "^0.11.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
@@ -45,6 +46,7 @@
     "lucide-react": "^0.465.0",
     "next": "^15.0.3",
     "next-intl": "^3.25.3",
+    "next-themes": "^0.4.4",
     "pg": "^8.13.1",
     "pino": "^9.5.0",
     "pino-pretty": "^13.0.0",
@@ -101,6 +103,7 @@
     "eslint-plugin-playwright": "^2.1.0",
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.14",
+    "eslint-plugin-simple-import-sort": "^12.1.1",
     "eslint-plugin-tailwindcss": "^3.17.5",
     "eslint-plugin-testing-library": "^7.0.0",
     "http-server": "^14.1.1",
@@ -113,6 +116,7 @@
     "semantic-release": "^24.2.0",
     "start-server-and-test": "^2.0.8",
     "storybook": "^8.4.6",
+    "storybook-dark-mode": "^4.0.2",
     "tailwindcss": "^3.4.15",
     "typescript": "^5.7.2",
     "vite-tsconfig-paths": "^5.1.3",
@@ -146,5 +150,8 @@
       "@semantic-release/git",
       "@semantic-release/github"
     ]
+  },
+  "volta": {
+    "node": "23.3.0"
   }
 }
