--- src/app/[locale]/(product-page)/product-page/page.tsx
@@ -1,28 +1,0 @@
-import ProductPage from '@/components/product-page';
-import { getTranslations, setRequestLocale } from 'next-intl/server';
-
-type IIndexProps = {
-  params: Promise<{ locale: string }>;
-};
-
-export async function generateMetadata(props: IIndexProps) {
-  const { locale } = await props.params;
-  const t = await getTranslations({
-    locale,
-    namespace: 'Index',
-  });
-
-  return {
-    title: t('meta_title'),
-    description: t('meta_description'),
-  };
-}
-
-export default async function Index(props: IIndexProps) {
-  const { locale } = await props.params;
-  setRequestLocale(locale);
-
-  return (
-    <ProductPage />
-  );
-};
