+++ src/app/[locale]/marketing/portfolio/[slug]/page.tsx
@@ -1,0 +1,72 @@
+import Image from 'next/image';
+import { getTranslations, setRequestLocale } from 'next-intl/server';
+
+import { routing } from '@/libs/i18nNavigation';
+
+type IPortfolioDetailProps = {
+  params: Promise<{ slug: string; locale: string }>;
+};
+
+export function generateStaticParams() {
+  return routing.locales
+    .map(locale =>
+      Array.from(Array.from({ length: 6 }).keys()).map(elt => ({
+        slug: `${elt}`,
+        locale,
+      })),
+    )
+    .flat(1);
+}
+
+export async function generateMetadata(props: IPortfolioDetailProps) {
+  const { locale, slug } = await props.params;
+  const t = await getTranslations({
+    locale,
+    namespace: 'PortfolioSlug',
+  });
+
+  return {
+    title: t('meta_title', { slug }),
+    description: t('meta_description', { slug }),
+  };
+}
+
+export default async function PortfolioDetail(props: IPortfolioDetailProps) {
+  const { locale, slug } = await props.params;
+  setRequestLocale(locale);
+  const t = await getTranslations({
+    locale,
+    namespace: 'PortfolioSlug',
+  });
+
+  return (
+    <>
+      <h1 className="capitalize">{t('header', { slug })}</h1>
+      <p>{t('content')}</p>
+
+      <div className="mt-5 text-center text-sm">
+        {`${t('log_management_powered_by')} `}
+        <a
+          className="text-blue-700 hover:border-b-2 hover:border-blue-700"
+          href="https://betterstack.com/?utm_source=github&utm_medium=sponsorship&utm_campaign=next-js-boilerplate"
+        >
+          Better Stack
+        </a>
+      </div>
+
+      <a
+        href="https://betterstack.com/?utm_source=github&utm_medium=sponsorship&utm_campaign=next-js-boilerplate"
+      >
+        <Image
+          className="mx-auto mt-2"
+          src="/assets/images/better-stack-dark.png"
+          alt="Better Stack"
+          width={128}
+          height={22}
+        />
+      </a>
+    </>
+  );
+};
+
+export const dynamicParams = false;
