+++ src/blocks/ProductGallery/ProductGallery.tsx
