+++ src/blocks/ProductTabs/ProductTabs.tsx
