+++ src/components/CounterForm.tsx
@@ -1,11 +1,12 @@
 'use client';
 
-import type { z } from 'zod';
-import { CounterValidation } from '@/validations/CounterValidation';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useTranslations } from 'next-intl';
 import { useRouter } from 'next/navigation';
+import { useTranslations } from 'next-intl';
 import { useForm } from 'react-hook-form';
+import type { z } from 'zod';
+
+import { CounterValidation } from '@/validations/CounterValidation';
 
 export const CounterForm = () => {
   const t = useTranslations('CounterForm');
