+++ src/components/CurrentCount.tsx
@@ -1,9 +1,10 @@
+import { eq } from 'drizzle-orm';
+import { headers } from 'next/headers';
+import { getTranslations } from 'next-intl/server';
+
 import { db } from '@/libs/DB';
 import { logger } from '@/libs/Logger';
 import { counterSchema } from '@/models/Schema';
-import { eq } from 'drizzle-orm';
-import { getTranslations } from 'next-intl/server';
-import { headers } from 'next/headers';
 
 export const CurrentCount = async () => {
   const t = await getTranslations('CurrentCount');
