+++ src/components/Hello.tsx
@@ -1,5 +1,6 @@
 import { currentUser } from '@clerk/nextjs/server';
 import { getTranslations } from 'next-intl/server';
+
 import { Sponsors } from './Sponsors';
 
 export const Hello = async () => {
