+++ src/components/ui/card.tsx
@@ -1,6 +1,6 @@
-import { cn } from '@/lib/utils';
-
 import * as React from 'react';
+
+import { cn } from '@/lib/utils';
 
 const Card = React.forwardRef<
   HTMLDivElement,
