+++ src/components/ui/input.tsx
@@ -1,6 +1,6 @@
-import { cn } from '@/lib/utils';
-
 import * as React from 'react';
+
+import { cn } from '@/lib/utils';
 
 const Input = React.forwardRef<HTMLInputElement, React.ComponentProps<'input'>>(
   ({ className, type, ...props }, ref) => {
