+++ src/templates/BaseTemplate.stories.tsx
@@ -1,7 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import messages from '@/locales/en.json';
 import { userEvent, within } from '@storybook/test';
 import { NextIntlClientProvider } from 'next-intl';
+
+import messages from '@/locales/en.json';
+
 import { BaseTemplate } from './BaseTemplate';
 
 const meta = {
