+++ src/templates/BaseTemplate.test.tsx
@@ -1,6 +1,8 @@
-import messages from '@/locales/en.json';
 import { render, screen, within } from '@testing-library/react';
 import { NextIntlClientProvider } from 'next-intl';
+
+import messages from '@/locales/en.json';
+
 import { BaseTemplate } from './BaseTemplate';
 
 describe('Base template', () => {
